added argon2 implementation as well, both can be toggled using keywords
