+++ frontend/src/utils/crypto/deriveKey.ts
@@ -1,0 +1,48 @@
+export const deriveKey = async (
+  password: string,
+  salt: Uint8Array,
+  method: "pbkdf2" | "argon2" = "pbkdf2"
+): Promise<CryptoKey> => {
+  if (method === "argon2") {
+    const argon2 = await import("argon2-browser");
+    const result = await argon2.hash({
+      pass: password,
+      salt,
+      type: argon2.ArgonType.Argon2id,
+      hashLen: 32,
+      time: 3,
+      mem: 65536,
+      parallelism: 1,
+    });
+
+    return await window.crypto.subtle.importKey(
+      "raw",
+      result.hash,
+      { name: "AES-GCM" },
+      false,
+      ["encrypt", "decrypt"]
+    );
+  } else {
+    const encoder = new TextEncoder();
+    const keyMaterial = await window.crypto.subtle.importKey(
+      "raw",
+      encoder.encode(password),
+      { name: "PBKDF2" },
+      false,
+      ["deriveKey"]
+    );
+
+    return await window.crypto.subtle.deriveKey(
+      {
+        name: "PBKDF2",
+        salt,
+        iterations: 100000,
+        hash: "SHA-256",
+      },
+      keyMaterial,
+      { name: "AES-GCM", length: 256 },
+      false,
+      ["encrypt", "decrypt"]
+    );
+  }
+};
